+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0b5] - Pending
+
+### Added
+
+- Added a service-uptime sensor to the DGB service device.
+
+### Changed
+
+- Report node uptime in seconds using the Home Assistant duration device class.
+
 ## [1.0.0b4] - 2026-08-30
 
 ### Fixed